requirements.txt
flask==3.1.3
flask-mail==0.10.0
psycopg2-binary==2.9.9
pdfkit==1.0.0
pyodbc==5.3.0
openpyxl==3.1.5
stripe==15.0.1
requests==2.33.1
qrcode[pil]==8.2
Pillow==12.2.0
opencv-python==4.13.0.90
pytesseract==0.3.13
ultralytics==8.4.33
python-dotenv==1.2.2
gunicorn==23.0.0
